requirements.txt
cs50
Flask
Flask-Session
requests
datetime, date
os
selenium
tempfile
werkzeug.exceptions
werkzeug.security
bs4
pandas
time